requirements.txt
camel-ai
pillow
pandas
streamlit
openai
agno
duckduckgo-search
yfinance
fastapi[standard] 
sqlalchemy
python-dotenv